[refactor] split into file constant
